feat: update response api config

+++ src/config/response.js
@@ -1,3 +1,4 @@
+import ENDPOINT from '@/config/endpoint'
 import router from '@/router/popup'
 import JSLib from '@/popup/services/services'
 const browserStorageService = JSLib.getBgService('storageService')()
@@ -9,21 +10,24 @@ export function handleResponseErrorMessage(error) {
       router.push({ name: 'Home' })
     }
     if (error.response.status === 401) {
-      browserStorageService.remove('cs_token')
-      storePromise.then((store) => {
-        store.commit('UPDATE_LOGIN_PAGE_INFO', null)
-        store.commit('UPDATE_IS_LOGGEDIN', false)
-        store.commit('CLEAR_ALL_DATA')
+      if (error.response.config.url !== ENDPOINT.SSO_AUTH) {
+        console.log('error', error.response);
+        browserStorageService.remove('cs_token')
+        storePromise.then((store) => {
+          store.commit('UPDATE_LOGIN_PAGE_INFO', null)
+          store.commit('UPDATE_IS_LOGGEDIN', false)
+          store.commit('CLEAR_ALL_DATA')
 
-        chrome.runtime.sendMessage({
-          command: 'updateStoreService',
-          sender: { key: 'isLoggedIn', value: false },
-        });
+          chrome.runtime.sendMessage({
+            command: 'updateStoreService',
+            sender: { key: 'isLoggedIn', value: false },
+          });
 
-        if (router.currentRoute.name !== 'login') {
-          router.push({ name: "login" });
-        }
-      })
+          if (router.currentRoute.name !== 'login') {
+            router.push({ name: "login" });
+          }
+        })
+      }
     }
   }
   return error
